pyproject.toml
[tool.poetry]
name = "classmanager"
version = "0.1.0"
description = ""
authors = ["Marilia Fernandez <marilia.mfernandez@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
djangorestframework = "^3.13.1"
psycopg2-binary = "^2.9.3"
python-decouple = "^3.6"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
